requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dj-database-url==2.2.0
inflection==0.5.1
PyMySQL==1.1.1
psycopg2-binary==2.9.10
PyJWT==2.11.0
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
whitenoise==6.8.2
